refactor: move MultiSelectForGroups to SemesterList
