+++ frontend/src/containers/SemesterPage/SemesterPage.js
@@ -4,15 +4,13 @@ import {
     getAllSemestersStart,
     getDisabledSemestersStart,
     getArchivedSemestersStart,
-    setGroupsToSemesterStart,
     handleSemesterStart,
     setError,
 } from '../../actions/semesters';
 import { fetchEnabledGroupsStart } from '../../actions/groups';
-import { setOpenErrorSnackbar, setOpenSuccessSnackbar } from '../../actions/snackbar';
+import { setOpenErrorSnackbar } from '../../actions/snackbar';
 
 const mapStateToProps = (state) => ({
-    semesters: state.semesters.semesters,
     archivedSemesters: state.semesters.archivedSemesters,
     isSnackbarOpen: state.snackbar.isSnackbarOpen,
     snackbarType: state.snackbar.snackbarType,
@@ -25,13 +23,10 @@ const mapDispatchToProps = (dispatch) => ({
     getAllSemestersItems: () => dispatch(getAllSemestersStart()),
     getDisabledSemestersItems: () => dispatch(getDisabledSemestersStart()),
     getArchivedSemestersItems: () => dispatch(getArchivedSemestersStart()),
-    setGroupsToSemester: (semesterId, groups) =>
-        dispatch(setGroupsToSemesterStart(semesterId, groups)),
     handleSemester: (values) => dispatch(handleSemesterStart(values)),
     setOpenErrorSnackbar: (message) => dispatch(setOpenErrorSnackbar(message)),
     setError: (res) => dispatch(setError(res)),
     getAllGroupsItems: () => dispatch(fetchEnabledGroupsStart()),
-    setOpenSuccessSnackbar: (message) => dispatch(setOpenSuccessSnackbar(message)),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(SemesterPage);
